KVM: TDX: Don't make TDH.MEM.RANGE.BLOCK if the key isn't configured
TDH.MEM.RANGE.BLOCK requires that the TD key is configured on the
hardware, so don't proceed to make the seamcall if the key isn't
configured. This can happen in the TD teardown flow, where the key is
freed before reclaiming memory pages. In this case, TDH.MEM.RANGE.BLOCK
is actually not needed as the TD isn't running (thus won't change the
sept mappings).

Signed-off-by: Wei Wang <wei.w.wang@intel.com>