KVM: TDX: add tdx cmd KVM_TDX_SET_MIGRATION_INFO
The KVM_TDX_SET_MIGRATION_INFO command is used by userspace to
set migration information. The information is passed via the
kvm_tdx_set_migration_info struct, and they will be stored in the MigTD
specific data area in the binding slot. The binding slot's state transits
to TDX_BINDING_SLOT_STATE_PREMIG_WAIT, indicating that the slot holds all
the data and waits for pre-migration to start. If the MigTD's vCPU is in
halt state, wake it up. MigTD will do TDG.VP.VMCALL to check the states of
the binding slots that have been added to it, and start pre-migration for
those who are in TDX_BINDING_SLOT_STATE_PREMIG_WAIT.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>