KVM: TDX: handle the ReportStatus cmd of TDG_VP_VMCALL_SERVICE.MIGTD
ReportStatus is a command of the MigTD service provided by the
TDG_VP_VMCALL_SERVICE function (defined in GHCI spec). It is
used by MigTD to report the status of the operation obtained via the
WaitForRequest command. For the "start migration" operation, it is for
MigTD to report back to KVM if the pre-migration is successfully done or
any error occurs (e.g. network issues detected by MigTD). If the status
is success, the related binding slot's state will be transit to
TDX_BINDING_SLOT_STATE_PREMIG_DONE, indicating pre-migation is done.
Later when migration is initiated, it can proceed to the migration phase
to export TD states. Upon errors reported in status, the slot's state
transits to TDX_BINDING_SLOT_STATE_BOUND. Userspace many fix the issues
and re-sent the migration info via KVM_TDX_SET_MIGRATION_INFO to re-start
the pre-migration step.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>