KVM: TDX: add tdx_mig_state to manage the migration related info of a TD
Add a per-TD scope struct, tdx_mig_state, to manage its migration
related info, e.g. the number of migration streams that have been
created in the TDX module. Some of the migration states are required
by the TDX architecture to be retained during the TD lifecycle (e.g.
migration streams that have been created) and re-used by different
migration sessions, so create the tdx_mig_state during the launch of
the TD if it's configured to be migratable and free it when destroy the
TD.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>