efi/unaccepted: Make sure unaccepted table is mapped in crashkernel case
Use memblock_add() to make sure that the table is mapped in direct
mapping.

During normal boot it happens automatically because the table is
allocated from usable memory. But during crashkernel boot only
memory specifically reserved for crash scenario is mapped.
memblock_add() forces the table to be mapped in crashkernel case.

Align the range to the nearest page boarders. Ranges smaller than page
size are not going to be mapped.

Signed-off-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>